feat: improved response time to microseconds
